Move grid related code to it's own field
