+++ JustBlog.UI/App/Blog/Post/CreatePostControllerBase.ts
@@ -5,7 +5,7 @@ import { NotificationFactory } from "Notification/notification.factory";
 import { Post } from "Blog/Post/Post";
 import { Tag } from "Blog/Tag/Tag";
 import { PostQueryFilter } from "Blog/Post/PostQueryFilter";
-import { GridQuery } from "Core/Models/GridQuery";
+import { GridQuery } from "Core/grid/GridQuery";
 
 export class CreatePostControllerBase extends BaseController {
