+++ JustBlog.UI/App/Core/grid/IQueryFilter.ts
