+++ JustBlog.UI/JustBlog.UI.csproj
@@ -64,7 +64,4 @@
   <ItemGroup>
     <WCFMetadata Include="Connected Services" />
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="App\Core\grid\" />
-  </ItemGroup>
 </Project>
